Homing

Approved-by: Shailesh Patil

+++ migrations/1608223534_create_piercing_table.up.sql
@@ -2,7 +2,7 @@ CREATE TYPE discard_type AS ENUM ('at_pickup_passing', 'at_discard_box');
 
 CREATE TABLE IF NOT EXISTS piercing(
  id uuid,
- cartridge_ids decimal[] NOT NULL,
+ cartridge_ids int[] NOT NULL,
  discard discard_type DEFAULT 'at_pickup_passing',
  created_at timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP,
  updated_at timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP,
